requirements.txt
pandas
streamlit
asyncio
google-generativeai
httpx
elementpath